Kracken-trading-bot-v2-main/kraken_bot_v300_infinity/legacy_v200E/kraken_ml_bot_v200E_full/requirements.txt
ccxt
aiohttp
fastapi
uvicorn
pydantic
python-dotenv
PyYAML
numpy
pandas
scipy
prometheus-client
jinja2
pytest
ruff
flake8
xgboost
scikit-learn
joblib
